refactor(jetbrains): move v5 migration under onboarding/providers/v5migration
Relocate the frontend 'migration' package to
ai.kilocode.client.onboarding.providers.v5migration so the v5 legacy
migration lives alongside the onboarding framework it now plugs into,
under a dedicated providers/ directory that future onboarding
providers can follow. Package-only rename — class names, RPC/backend
code, and DTOs are unchanged.
