+++ packages/kilo-jetbrains/frontend/src/test/kotlin/ai/kilocode/client/onboarding/providers/v5migration/MigrationOnboardingProviderTest.kt
@@ -1,4 +1,4 @@
-package ai.kilocode.client.migration
+package ai.kilocode.client.onboarding.providers.v5migration
 
 import ai.kilocode.rpc.dto.LegacyMigrationDetectionDto
 import com.intellij.testFramework.fixtures.BasePlatformTestCase
